Modified response generation, added qualification selector
The most difficult response to add is the qualification selector. As
I want it to ask the user for input. This is implemented, and when
the user selects a button it nicely pops text in on behalf of the user.

This immediatly broke the whole response generation logic. So I bit the
bullet and implemented it as redux middleware which feels slightly
dirty, but does work in the most general case.