Added technologies section, revamped response factory
The response factory was just a big if/else statement which was
really hard to read. Now it's a series of functions that feeds
a `reduce()` call which matches the text and outputs any responses.

A new technologies response has been added as well. This responds
to the qualification selector output (IE: `I am a seasoned developer`)
and shows the qualifications for that particular user type.