Created generalized rendering system
After building out a few views I came to realize that most of this
stuff was pretty much boilerplate. I was also pretty tightly
coupling the view to the underlying data, which isn't really
neccessary.

It turns out the data is in a pretty specific format and that by
implementing just a few rules I could reliably render the data
without having to write out each view manually.

The new `response-renderers.js` is the result. It takes an object
and runs through it producing a DOM that represents that data.
I can now churn out the remaining views very quickly as they all
work using basically the same language I've already implemented.